requirements.txt
PyMuPDF
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
transformers
torch
numpy
Pillow
python-dotenv
faiss-cpu